* Fix problem with self-signed SSL certificates. * Improve Avahi/Zeroconf configuration scripts.
!!! Note, this f**king software ios too stupid to remember an array of trusted_domains. After 'omv-mkconf owncloud' the config.php looks like

  'trusted_domains' => array (
    0 => "dhcppc3:8484",
    1 => "dhcppc3.local:8484",
  ),

but after login using one of those domains the list is shortened to the previously used one. This does not make fun anymore. Who the hell is using this xyz *$? software. I can't imagine how it is possible to catch up millions of dollars startup capital for such a software.

git-svn-id: svn://svn.code.sf.net/p/openmediavault/code@1298 3c3d7b80-5828-415e-9f61-280b98178fff